[modify] java / USER 실습 내용 수정 및 변경 - Specifications 추가

+++ sesac/data/java/user/dto/UserResoponseDto.java
@@ -9,8 +9,7 @@ import java.time.LocalDateTime;
 @Getter
 @Builder
 public class UserResoponseDto {
-    private final long id;
-
+    private final int id;
     private final String username;
     private final String email;
     private final String nickname;
